+++ packages/bpk-component-calendar/stories.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { storiesOf, action } from '@kadira/storybook';
-import isWeekend from 'date-fns/is_weekend';
 import addMonths from 'date-fns/add_months';
 import {
   formatMonth,
@@ -15,7 +14,15 @@ import {
   weekDays,
 } from './test-utils';
 
-import BpkCalendar, { BpkCalendarView, BpkCalendarGrid, BpkCalendarNav, BpkCalendarDate } from './index';
+import BpkCalendar, {
+  // BpkCalendarView,
+  BpkCalendarGrid,
+  BpkCalendarGridHeader,
+  BpkCalendarNav,
+  BpkCalendarDate,
+} from './index';
+
+import MonthViewCalendar from './stories-components';
 
 class CalendarContainer extends Component {
   constructor() {
@@ -30,7 +37,11 @@ class CalendarContainer extends Component {
       <BpkCalendar
         {...this.props}
         date={this.state.date}
-        onDateSelect={date => this.setState({ date, focusedDate: date })}
+        onDateSelect={(date) => {
+          this.setState({ date, focusedDate: date });
+          action('Selected day')(date);
+        }}
+        onMonthChange={action('Changed month')}
       />
     );
   }
@@ -41,11 +52,18 @@ storiesOf('bpk-component-calendar', module)
     <BpkCalendarNav
       month={new Date()}
       changeMonthLabel="Change month"
-      onChangeMonth={action('Changed month')}
+      onMonthChange={action('Changed month')}
       minDate={new Date()}
       maxDate={addMonths(new Date(), 12)}
       formatMonth={formatMonth}
       id="myCalendarNav"
+    />
+  ))
+  .add('BpkCalendarGridHeader', () => (
+    <BpkCalendarGridHeader
+      weekStartsOn={1}
+      daysOfWeek={weekDays}
+      showWeekendSeparator
     />
   ))
   .add('BpkCalendarGrid', () => (
@@ -61,21 +79,6 @@ storiesOf('bpk-component-calendar', module)
       preventKeyboardFocus
     />
   ))
-  .add('BpkCalendarView', () => (
-    <BpkCalendarView
-      month={new Date()}
-      formatDateFull={formatDateFull}
-      DateComponent={BpkCalendarDate}
-      formatMonth={formatMonth}
-      minDate={new Date()}
-      maxDate={addMonths(new Date(), 12)}
-      daysOfWeek={weekDays}
-      changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
-      id="myCalendar"
-      preventKeyboardFocus
-    />
-  ))
   .add('Calendar - default', () => (
     <CalendarContainer
       id="myCalendar"
@@ -83,7 +86,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDays}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
     />
   ))
   .add('Calendar - Don\'t show weekend separator', () => (
@@ -93,7 +95,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDays}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       showWeekendSeparator={false}
     />
   ))
@@ -104,7 +105,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDays}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       weekStartsOn={0}
     />
   ))
@@ -115,7 +115,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDaysMoreWeekend}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
     />
   ))
   .add('Calendar - ar-AE locale', () => (
@@ -125,7 +124,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFullArabic}
       daysOfWeek={weekDaysArabic}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       weekStartsOn={6}
     />
   ))
@@ -136,7 +134,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFullJapanese}
       daysOfWeek={weekDaysJapanese}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       weekStartsOn={0}
     />
   ))
@@ -147,7 +144,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDays}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       minDate={new Date(2020, 4, 15)}
       maxDate={new Date(2020, 5, 15)}
     />
@@ -159,7 +155,6 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDays}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       markToday={false}
     />
   ))
@@ -170,34 +165,9 @@ storiesOf('bpk-component-calendar', module)
       formatDateFull={formatDateFull}
       daysOfWeek={weekDays}
       changeMonthLabel="Change month"
-      onDateSelect={action('Selected day')}
       markOutsideDays={false}
     />
   ))
-  .add('Calendar -  Custom date component', () => {
-    const MyCustomDate = (props) => {
-      const cx = {
-        backgroundColor: 'rgb(255, 84, 82)',
-        width: '50%',
-        height: '50%',
-        borderRadius: '5rem',
-        margin: '25%',
-      };
-      if (isWeekend(props.date)) { cx.backgroundColor = 'rgb(0, 215, 117)'; }
-      return <div style={cx} />;
-    };
-    MyCustomDate.propTypes = {
-      date: React.PropTypes.instanceOf(Date).isRequired,
-    };
-    return (
-      <CalendarContainer
-        id="myCalendar"
-        formatMonth={formatMonth}
-        formatDateFull={formatDateFull}
-        daysOfWeek={weekDays}
-        changeMonthLabel="Change month"
-        onDateSelect={action('Selected day')}
-        DateComponent={MyCustomDate}
-      />
-    );
-  });
+  .add('Custom composed calendar', () => (
+    <MonthViewCalendar />
+  ));