+++ changelog.md
@@ -2,7 +2,12 @@
 
 ## UNRELEASED
 
-_Nothing yet..._
+**Changed:**
+- bpk-component-calendar:
+  - `onMonthChange` callback added
+  - BpkCalendarView has been replaced by the higher-order component `composeCalendar`
+  - Higher-order component `withCalendarState` has been added to provide focus management and keyboard input
+  - Renamed the `date` prop to `selectedDate`
 
 ## 2017-05-01 - New form validation & fieldset components